RedfishClientPkg/ConvertLib: Redfish JSON schema to C structure source files
Total 7045 files that is unable to be reviewed.

The Redfish JSON schema to C structure source and header files
are auto-generated by the open source Redfish Schema C Structure
Generator Python script.
(https://github.com/DMTF/Redfish-Schema-C-Struct-Generator).
Each pair of the C source file and header file are generated base
on the Redfish schema name space and resource type, and can be built as the native C library.
The files under ConvertLib/include and ConvertLib/src will be
updated against to the Redfish schema release.
(this commit is based on Redfish schema 2020.4 release)

Signed-off-by: Abner Chang <abner.chang@hpe.com>